refactor: move methods to own files
